Extract the clipboard write shared by the copy features
copy-markdown.js carries a secure-context check plus an execCommand fallback
for browsers where navigator.clipboard is unavailable. The heading copy button
needs the same logic, so it moves to its own module rather than becoming a
second copy.

Behavior is unchanged: the same two code paths in the same order, with the
warning prefix generalized from [copy-md] to [clipboard].

code-blocks.js still has its own simpler inline version. It is bound up in a
larger icon-swapping flow, so it is left alone.
